Add README note re: thread count global variable
I'd like to suggest adding this note to the README.

I've found myself digging into the source code a couple times to re-assign the thread count.  Seems like a good idea to have it in README since this has been a repeated question for me :+1:.

Btw, this is a very handy gem for creating a quick & dirty threaded solution (I've been using it with web requests).