Uses the size of the Default ThreadGroup to check thread limits
Instead of using sleep and timing to check if threading is occuring,
we can instead check the number of threads in the default thread
group. It should always be the thread limit plus 1 for the main
thread. We obtain the size of the thread group on the last
element in the enumerable and check it to see if it meets the
expectations.

This test will break if we decide to use a different ThreadGroup
to manage the threads spawned in the ThreadPool and we'd be back
to square one.

This test would also break if the thread spawning algorithm is
optimized further.