Passes tests that don't pass a block to #pmap
With the normal #map method, the documentation states "If no block
is given, an enumerator is returned instead". Based on this premise,
self can be returned when a block isn't given because self is
an enumerator.

For the test to pass, we must compare the results of enumeration
with both pmap and map. Comparing the enumerators themselves may
yield different classes. Therefore #to_a is compared, which
effectively normalizes the results of enumeration.